+++ griptape/drivers/vector/azure_mongodb_vector_store_driver.py
@@ -16,7 +16,7 @@ class AzureMongoDbVectorStoreDriver(MongoDbAtlasVectorStoreDriver):
         include_vectors: bool = False,
         offset: Optional[int] = None,
         **kwargs,
-    ) -> list[BaseVectorStoreDriver.QueryResult]:
+    ) -> list[BaseVectorStoreDriver.Entry]:
         """Queries the MongoDB collection for documents that match the provided query string.
 
         Results can be customized based on parameters like count, namespace, inclusion of vectors, offset, and index.
@@ -50,7 +50,7 @@ class AzureMongoDbVectorStoreDriver(MongoDbAtlasVectorStoreDriver):
         pipeline.append({"$project": {"similarityScore": {"$meta": "searchScore"}, "document": "$$ROOT"}})
 
         return [
-            BaseVectorStoreDriver.QueryResult(
+            BaseVectorStoreDriver.Entry(
                 id=str(doc["_id"]),
                 vector=doc[self.vector_path] if include_vectors else [],
                 score=doc["similarityScore"],