+++ griptape/drivers/vector/mongodb_atlas_vector_store_driver.py
@@ -110,7 +110,7 @@ class MongoDbAtlasVectorStoreDriver(BaseVectorStoreDriver):
         include_vectors: bool = False,
         offset: Optional[int] = None,
         **kwargs,
-    ) -> list[BaseVectorStoreDriver.QueryResult]:
+    ) -> list[BaseVectorStoreDriver.Entry]:
         """Queries the MongoDB collection for documents that match the provided query string.
 
         Results can be customized based on parameters like count, namespace, inclusion of vectors, offset, and index.
@@ -148,7 +148,7 @@ class MongoDbAtlasVectorStoreDriver(BaseVectorStoreDriver):
             pipeline[0]["$vectorSearch"]["filter"] = {"namespace": namespace}
 
         results = [
-            BaseVectorStoreDriver.QueryResult(
+            BaseVectorStoreDriver.Entry(
                 id=str(doc["_id"]),
                 vector=doc[self.vector_path] if include_vectors else [],
                 score=doc["score"],