+++ griptape/drivers/vector/redis_vector_store_driver.py
@@ -108,13 +108,13 @@ class RedisVectorStoreDriver(BaseVectorStoreDriver):
         namespace: Optional[str] = None,
         include_vectors: bool = False,
         **kwargs,
-    ) -> list[BaseVectorStoreDriver.QueryResult]:
+    ) -> list[BaseVectorStoreDriver.Entry]:
         """Performs a nearest neighbor search on Redis to find vectors similar to the provided input vector.
 
         Results can be limited using the count parameter and optionally filtered by a namespace.
 
         Returns:
-            A list of BaseVectorStoreDriver.QueryResult objects, each encapsulating the retrieved vector, its similarity score, metadata, and namespace.
+            A list of BaseVectorStoreDriver.Entry objects, each encapsulating the retrieved vector, its similarity score, metadata, and namespace.
         """
         Query = import_optional_dependency("redis.commands.search.query").Query
 
@@ -140,7 +140,7 @@ class RedisVectorStoreDriver(BaseVectorStoreDriver):
             vector_id = document.id.split(":")[1] if ":" in document.id else document.id
             vector_float_list = json.loads(document.vec_string) if include_vectors else None
             query_results.append(
-                BaseVectorStoreDriver.QueryResult(
+                BaseVectorStoreDriver.Entry(
                     id=vector_id,
                     vector=vector_float_list,
                     score=float(document.score),